MySystem/requirements.txt
asgiref==3.3.1
Django==3.1.3
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
future==0.18.2
numpy==1.19.4
pandas==1.1.4
PyJWT==1.7.1
PyMySQL==0.10.1
python-dateutil==2.8.1
pytz==2020.4
regex==2020.11.13
serializers==0.2.4
six==1.15.0
sqlparse==0.4.1
